refactor: 1단계 리뷰 반영
